[*] port to MaxOSX

+++ src/axl_gui/axl_gui_ImageDesc.cpp
@@ -15,7 +15,7 @@ getPixelFormatString (PixelFormat pixelFormat)
 		"RGB",  // PixelFormat_Rgb,
 	};
 
-	return pixelFormat < countof (stringTable) ? 
+	return (size_t) pixelFormat < countof (stringTable) ?
 		stringTable [pixelFormat] : 
 		"undefined-pixel-format";
 }
